requirements.txt
Flask
pandas
scikit-learn
xgboost
requests
psycopg2-binary
